+++ public/chat_header_button.tsx
@@ -9,8 +9,6 @@ import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useEffectOnce, useObservable } from 'react-use';
 
 import { ApplicationStart, HeaderVariant, SIDECAR_DOCKED_MODE } from '../../../src/core/public';
-// TODO: Replace with getChrome().logos.Chat.url
-import chatIcon from './assets/chat.svg';
 import { getIncontextInsightRegistry } from './services';
 import { ChatFlyout } from './chat_flyout';
 import { ChatContext, IChatContext } from './contexts/chat_context';
@@ -26,6 +24,7 @@ import {
 import { useCore } from './contexts/core_context';
 import { MountPointPortal } from '../../../src/plugins/opensearch_dashboards_react/public';
 import { usePatchFixedStyle } from './hooks/use_patch_fixed_style';
+import { getLogoIcon } from './services';
 
 interface HeaderChatButtonProps {
   application: ApplicationStart;
@@ -36,11 +35,10 @@ interface HeaderChatButtonProps {
   inLegacyHeader?: boolean;
 }
 
-let flyoutLoaded = false;
-
 export const HeaderChatButton = (props: HeaderChatButtonProps) => {
   const core = useCore();
   const { inLegacyHeader } = props;
+  const sideCarRef = useRef<{ close: Function }>();
   const [appId, setAppId] = useState<string>();
   const [conversationId, setConversationId] = useState<string>();
   const [title, setTitle] = useState<string>();
@@ -51,6 +49,9 @@ export const HeaderChatButton = (props: HeaderChatButtonProps) => {
   const [interactionId, setInteractionId] = useState<string | undefined>(undefined);
   const [inputFocus, setInputFocus] = useState(false);
   const inputRef = useRef<HTMLInputElement>(null);
+  const flyoutLoadedRef = useRef(false);
+  const flyoutVisibleRef = useRef(flyoutVisible);
+  flyoutVisibleRef.current = flyoutVisible;
   const registry = getIncontextInsightRegistry();
   const headerVariant = useObservable(core.services.chrome.getHeaderVariant$());
   const isSingleLineHeader = headerVariant === HeaderVariant.APPLICATION;
@@ -131,24 +132,25 @@ export const HeaderChatButton = (props: HeaderChatButtonProps) => {
   };
 
   useEffect(() => {
-    if (!flyoutLoaded && flyoutVisible) {
+    if (!flyoutLoadedRef.current && flyoutVisible) {
       const mountPoint = flyoutMountPoint.current;
       if (mountPoint) {
-        core.overlays.sidecar().open(mountPoint, {
+        sideCarRef.current = core.overlays.sidecar().open(mountPoint, {
           className: 'chatbot-sidecar',
           config: {
             dockedMode: SIDECAR_DOCKED_MODE.RIGHT,
             paddingSize: DEFAULT_SIDECAR_LEFT_OR_RIGHT_SIZE,
+            isHidden: false,
           },
         });
-        flyoutLoaded = true;
+        flyoutLoadedRef.current = true;
       }
-    } else if (flyoutLoaded && flyoutVisible) {
+    } else if (flyoutLoadedRef.current && flyoutVisible) {
       core.overlays.sidecar().show();
-    } else if (flyoutLoaded && !flyoutVisible) {
+    } else if (flyoutLoadedRef.current && !flyoutVisible) {
       core.overlays.sidecar().hide();
     }
-  }, [flyoutVisible, flyoutLoaded]);
+  }, [flyoutVisible]);
 
   const onKeyUp = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === 'Escape') {
@@ -170,6 +172,15 @@ export const HeaderChatButton = (props: HeaderChatButtonProps) => {
 
     return () => {
       document.removeEventListener('keydown', onGlobalMouseUp);
+    };
+  }, []);
+
+  useEffect(() => {
+    return () => {
+      if (flyoutVisibleRef.current) {
+        core.overlays.sidecar().hide();
+      }
+      sideCarRef.current?.close();
     };
   }, []);
 
@@ -223,7 +234,7 @@ export const HeaderChatButton = (props: HeaderChatButtonProps) => {
       {!inLegacyHeader && isSingleLineHeader && (
         <EuiButtonIcon
           className={classNames(['eui-hideFor--xl', 'eui-hideFor--xxl', 'eui-hideFor--xxxl'])}
-          iconType={chatIcon}
+          iconType={getLogoIcon('gray')}
           onClick={() => setFlyoutVisible(!flyoutVisible)}
           display="base"
           size="s"
@@ -253,7 +264,7 @@ export const HeaderChatButton = (props: HeaderChatButtonProps) => {
         />
         <EuiIcon
           aria-label="toggle chat flyout icon"
-          type={chatIcon}
+          type={getLogoIcon('gray')}
           size="m"
           onClick={() => setFlyoutVisible(!flyoutVisible)}
           className="llm-chat-toggle-icon"